Replace binary entirely
