+++ offload/tools/CMakeLists.txt
@@ -14,4 +14,3 @@ endmacro()
 
 add_subdirectory(deviceinfo)
 add_subdirectory(kernelreplay)
-add_subdirectory(liboffload-deviceinfo)
